requirements.txt
click==8.1.4
coloredlogs==15.0.1
Flask==2.2.5
Flask-Cors==4.0.0
flatbuffers==23.5.26
humanfriendly==10.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
numpy
tqdm==4.64.1
pandas
gunicorn
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
onnxruntime==1.14.1
opencv-python==4.8.0.74
packaging==23.1
Pillow==9.5.0
protobuf==4.23.4
sympy==1.10.1
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
vietocr
gradio
